Adjust code to fit the PR comments
