+++ src/main/java/org/elasql/procedure/calvin/CalvinStoredProcedure.java
@@ -35,7 +35,6 @@ import org.elasql.server.Elasql;
 import org.elasql.sql.PrimaryKey;
 import org.elasql.storage.metadata.NotificationPartitionPlan;
 import org.elasql.storage.tx.concurrency.ConservativeOrderedCcMgr;
-import org.elasql.storage.tx.concurrency.fifolocker.FifoLock;
 import org.elasql.storage.tx.recovery.DdRecoveryMgr;
 import org.vanilladb.core.remote.storedprocedure.SpResultSet;
 import org.vanilladb.core.sql.Constant;
@@ -54,7 +53,6 @@ public abstract class CalvinStoredProcedure<H extends StoredProcedureParamHelper
 
 	private ExecutionPlan execPlan;
 	private Transaction tx;
-	private FifoLock myFifoLock = new FifoLock(txNum);
 	private boolean isCommitted = false;
 
 	public CalvinStoredProcedure(long txNum, H paramHelper) {
