+++ src/main/java/org/elasql/procedure/naive/NaiveStoredProcedure.java
@@ -25,7 +25,6 @@ import org.elasql.cache.naive.NaiveCacheMgr;
 import org.elasql.server.Elasql;
 import org.elasql.sql.PrimaryKey;
 import org.elasql.storage.tx.concurrency.ConservativeOrderedCcMgr;
-import org.elasql.storage.tx.concurrency.fifolocker.FifoLock;
 import org.elasql.storage.tx.recovery.DdRecoveryMgr;
 import org.vanilladb.core.remote.storedprocedure.SpResultSet;
 import org.vanilladb.core.sql.Constant;
@@ -44,7 +43,6 @@ public abstract class NaiveStoredProcedure<H extends StoredProcedureParamHelper>
 	// Record keys
 	private List<PrimaryKey> readKeys = new ArrayList<PrimaryKey>();
 	private List<PrimaryKey> writeKeys = new ArrayList<PrimaryKey>();
-	private FifoLock myFifoLock = new FifoLock(txNum);
 	private boolean isCommitted = false; 
 	
 	private NaiveCacheMgr cacheMgr = (NaiveCacheMgr) Elasql.remoteRecReceiver();
